HW 2 done.

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -1,10 +1,14 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
+import ru.javawebinar.topjava.util.DateTimeUtil;
 import ru.javawebinar.topjava.util.MealsUtil;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -13,6 +17,8 @@ import java.util.concurrent.atomic.AtomicInteger;
  * GKislin
  * 15.09.2015.
  */
+
+@Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
@@ -23,25 +29,31 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
 
     @Override
     public Meal save(Meal meal) {
-        if (meal.isNew()) {
+        if (meal.isNew())
             meal.setId(counter.incrementAndGet());
-        }
         repository.put(meal.getId(), meal);
         return meal;
     }
 
     @Override
-    public void delete(int id) {
+    public boolean delete(int id) {
+        if (repository.get(id) == null)
+            return false;
         repository.remove(id);
+        return true;
     }
 
     @Override
     public Meal get(int id) {
+        if (repository.get(id) == null)
+            return null;
         return repository.get(id);
     }
 
     @Override
     public Collection<Meal> getAll() {
+        ArrayList<Meal> sortedMeal = new ArrayList<>(repository.values());
+        sortedMeal.sort(Comparator.comparing(Meal::getDate));
         return repository.values();
     }
 }
